Race and Median Income in St. Louis by Census Tract, 2017

A simple features data set containing the geometry and associated attributes for the 2013-2017 American Community Survey estimates for median household income and the percentage of white residents in St. Louis.

data(stl_race_income)

Format

A data frame with 106 rows and 4 variables:

GEOID

full GEOID string

pctWhite

Percent of white residents per tract

medInc

Median household income of tract

geometry

simple features geometry

Source

tidycensus package

Examples

str(stl_race_income)

#> Classes ‘sf’ and 'data.frame':	106 obs. of  4 variables:
#>  $ GEOID   : chr  "29510112100" "29510116500" "29510110300" "29510103700" ...
#>  $ pctWhite: num  66.47 47.8 2.43 90.77 87.73 ...
#>  $ medInc  : num  56118 43913 17448 50565 74425 ...
#>  $ geometry:sfc_POLYGON of length 106; first list element: List of 1
#>   ..$ : num [1:284, 1:2] -90.3 -90.3 -90.3 -90.3 -90.3 ...
#>   ..- attr(*, "class")= chr [1:3] "XY" "POLYGON" "sfg"
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA
#>   ..- attr(*, "names")= chr [1:3] "GEOID" "pctWhite" "medInc"

head(stl_race_income)

#> Simple feature collection with 6 features and 3 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: -90.32052 ymin: 38.59208 xmax: -90.22502 ymax: 38.67066
#> CRS:            EPSG:4269
#>         GEOID  pctWhite medInc                       geometry
#> 1 29510112100 66.470727  56118 POLYGON ((-90.30445 38.6328...
#> 2 29510116500 47.799564  43913 POLYGON ((-90.24302 38.5975...
#> 3 29510110300  2.427686  17448 POLYGON ((-90.24032 38.6643...
#> 4 29510103700 90.773067  50565 POLYGON ((-90.29877 38.6028...
#> 5 29510103800 87.733402  74425 POLYGON ((-90.32052 38.5941...
#> 6 29510104500 74.723618  54286 POLYGON ((-90.29432 38.6209...

summary(stl_race_income$medInc)

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   10545   23474   34688   37140   51379   74425