Race and Median Income in St. Louis by Census Tract, 2017

A simple features data set containing the geometry and associated attributes for the 2013-2017 American Community Survey estimates for median household income and the percentage of white residents in St. Louis. This version of the sample data are stored as point data.

data(stl_race_income_point)

Format

A data frame with 106 rows and 4 variables:

GEOID

full GEOID string

pctWhite

Percent of white residents per tract

medInc

Median household income of tract

geometry

simple features geometry

Source

tidycensus package

Examples

str(stl_race_income_point)
#> Classes ‘sf’ and 'data.frame':	106 obs. of  4 variables:
#>  $ GEOID   : chr  "29510102300" "29510102400" "29510104500" "29510106100" ...
#>  $ pctWhite: num  82.93 90.12 74.72 1.75 2.38 ...
#>  $ medInc  : num  51650 45375 54286 18895 36130 ...
#>  $ geometry:sfc_POINT of length 106; first list element:  'XY' num  -90.3 38.6
#>  - attr(*, "sf_column")= chr "geometry"
#>  - attr(*, "agr")= Factor w/ 3 levels "constant","aggregate",..: NA NA NA
#>   ..- attr(*, "names")= chr [1:3] "GEOID" "pctWhite" "medInc"
head(stl_race_income_point)
#> Simple feature collection with 6 features and 3 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -90.28881 ymin: 38.56287 xmax: -90.27698 ymax: 38.6694
#> Geodetic CRS:  NAD83
#>         GEOID  pctWhite medInc                   geometry
#> 1 29510102300 82.925682  51650 POINT (-90.28698 38.56287)
#> 2 29510102400 90.116046  45375 POINT (-90.27785 38.57716)
#> 3 29510104500 74.723618  54286 POINT (-90.28087 38.62656)
#> 4 29510106100  1.752464  18895  POINT (-90.28116 38.6694)
#> 5 29510105500  2.376729  36130 POINT (-90.27698 38.66003)
#> 6 29510105200 36.833277  60938 POINT (-90.28881 38.65056)
summary(stl_race_income_point$medInc)
#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>   10545   23474   34688   37140   51379   74425