Apply Bivariate Color to ggplot Object

Applies the selected palette as the color aesthetic when geom_sf is used and the bi_class variable is given as the color in the aesthetic mapping.

bi_scale_color(pal, dim = 3, flip_axes = FALSE, rotate_pal = FALSE, ...)

Arguments

pal

A palette name; one of "Bluegill", "BlueGold", "BlueOr", "BlueYl", "Brown"/"Brown2", "DkBlue"/"DkBlue2", "DkCyan"/"DkCyan2", "DkViolet"/"DkViolet2", "GrPink"/"GrPink2", "PinkGrn", "PurpleGrn", or "PurpleOr".

Pairs of palettes, such as "GrPink"/"GrPink2", are included for legacy support. The numbered palettes support four-by-four bivarite maps while the un-numbered ones, which were the five included in the original release of the package, only support two-by-two and three-by-three maps.

dim

The dimensions of the palette, either 2 for a two-by-two palette, 3 for a three-by-three palette, or 4 for a four-by-four palette.

flip_axes

A logical scalar; if TRUE the axes of the palette will be flipped. If FALSE (default), the palette will be displayed on its original axes.

rotate_pal

A logical scalar; if TRUE the palette will be rotated 180 degrees. If FALSE (default), the palette will be displayed in its original orientation

...

Arguments to pass to scale_color_manual

Value

A ggplot object with the given bivariate palette applied to the data.